Core: Update to 10.1
Port From (https://github.com/TrinityCore/TrinityCore/commit/16bc74667e8996b64258718e95b97258dfc0217a)

+++ Source/Game/DataStorage/Structs/U_Records.cs
@@ -12,7 +12,7 @@ namespace Game.DataStorage
         public uint Id;
         public int ParentUiMapID;
         public int Flags;
-        public uint System;
+        public byte System;
         public UiMapType Type;
         public int BountySetID;
         public uint BountyDisplayLocation;
