+++ Source/Game/DataStorage/Structs/V_Records.cs
@@ -11,7 +11,7 @@ namespace Game.DataStorage
     {
         public uint Id;
         public VehicleFlags Flags;
-        public byte FlagsB;
+        public int FlagsB;
         public float TurnSpeed;
         public float PitchSpeed;
         public float PitchMin;
